refactor: 불필요한 Suspense wrapping
